Create How to Bring a Health Lens to Environmental Monitoring?
Hi, 
I’m Celia Santos Tapia from ISGlobal. I’ve put together this entry about how to include or take into account the health perspective in monitoring campaigns. I’m not sure if the format is right (especially the links) or if this is the best place for it. I’d also like to know if the content makes sense for the toolkit. 

Anyway, here it is — let me know what you think! Cheers!